Store settings,cache and logs in central user directory
