+++ core/__init__.py
@@ -1,9 +1,8 @@
-import logging.config
-
 from .cache import CourseCache
 from .driver_manager import ALL_VALID_BROWSER_STRINGS, DriverManager
+from .logging import load_logging_config
 from .scrapers.manager import ScraperManager
 from .settings import Settings
 from .udemy import UdemyActions
 
-logging.config.fileConfig("logconfig.ini", disable_existing_loggers=False)
+load_logging_config()
