+++ logconfig.ini
@@ -19,7 +19,7 @@ qualname=udemy_enroller
 propagate=0
 
 [handler_defaultHandler]
-class=FileHandler
+class=core.logging.CustomFileHandler
 formatter=defaultFormatter
 args=("app.log", "a")
